Add very basic Markdown parser
This is all we need to get Markdown text to be displayed in the editor. We're not interested in parsing all of MD or parsing it into HTML. We want to be able do style only some very limited parts of it and this is all we need.

Also some cleanup to remove my previous MD display attempts